--- App.css
@@ -1,77 +1,0 @@
-.app-container {
-  display: flex;
-  flex-direction: column;
-  /* height: 100vh; */
-  min-width: 100%;
-  width: 100%;
-  max-width: 100%;
-  min-height: 100%;
-  height: 100%;
-  max-height: 100%;
-  top: 0;
-  left: 0;
-  padding: 0px;
-  margin: 0px;
-  z-index: 1;
-}
-
-.dark {
-  background-color: #333;
-  color: #fff;
-}
-
-.light {
-  background-color: #fff;
-  color: #333;
-}
-
-.header {
-  display: flex;
-  padding: 1rem;
-  align-items: center;
-  background-color: inherit;
-  justify-content: space-between;
-  position: sticky;
-  top: 0;
-  overflow: scroll;
-  overflow-y: scroll;
-}
-
-.settings-button {
-  cursor: pointer;
-  align-self: flex-start;
-  padding: 0;
-  background-color: #333
-}
-
-.main-container {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  padding: 1rem;
-  height: 100%;
-}
-
-.bible-menu {
-  margin-bottom: 1rem;
-}
-
-.bible-text {
-  flex: 1;
-  overflow-y: auto;
-  font-size: 1.25rem;
-  line-height: 1.5;
-}
-
-.bible-text p {
-  margin: 0;
-}
-
-.bible-text .verse-number {
-  font-weight: bold;
-  margin-right: 0.5rem;
-}
-
-.bible-text .verse-text {
-  display: inline;
-}
