--- assets/locales/de.json
@@ -1,6 +1,0 @@
-{
-    "select_language": "Wähle Sprache",
-    "select_font_size": "Wähle Schriftgröße",
-    "chapter": "Kapitel",
-    "verse": "Vers"
-}
