--- assets/locales/en.json
@@ -1,6 +1,0 @@
-{
-    "select_language": "Select Language",
-    "select_font_size": "Select Font Size",
-    "chapter": "Chapter",
-    "verse": "Verse"
-}
