--- components/BibleMenu.css
@@ -1,32 +1,0 @@
-.bible-menu {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  padding-top: 1rem;
-  align-self: flex-start;
-  align-items: flex-start;
-}
-
-#growth {
-  overflow-y: scroll;
-}
-
-.bible-menu button {
-  margin: 1rem;
-  margin-left: 0;
-  padding: 0.5rem;
-  border: none;
-  border-radius: 0rem;
-  background-color: #f5f5f5;
-  color: #333;
-  font-size: 1rem;
-  font-weight: bold;
-  cursor: pointer;
-  transition: background-color 0.2s ease-in-out;
-}
-
-.bible-menu button:hover {
-  background-color: #e0e0e0;
-}
