--- components/SettingsModal.css
@@ -1,127 +1,0 @@
-.settings-modal {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 9999;
-    background-color: rgba(0, 0, 0, 0.5);
-    visibility: hidden;
-    opacity: 0;
-    transition: visibility 0s, opacity 0.5s ease-in-out;
-  }
-  
-  .settings-modal.is-active {
-    visibility: visible;
-    opacity: 1;
-  }
-  
-  .settings-modal-background {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-  }
-  
-  .settings-modal-card {
-    position: relative;
-    width: 500px;
-    max-width: 90%;
-    background-color: #fff;
-    border-radius: 5px;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
-  }
-  
-  .settings-modal-card-head {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem;
-    border-bottom: 1px solid #ccc;
-  }
-  
-  .settings-modal-card-title {
-    font-size: 1.5rem;
-    font-weight: bold;
-  }
-  
-  .settings-modal-card-body {
-    padding: 1rem;
-  }
-  
-  .settings-modal-card-foot {
-    display: flex;
-    justify-content: flex-end;
-    padding: 1rem;
-    border-top: 1px solid #ccc;
-  }
-  
-  .label {
-    display: block;
-    margin-bottom: 0.5rem;
-    font-weight: bold;
-  }
-  
-  .control {
-    display: flex;
-    align-items: center;
-  }
-  
-  .select {
-    position: relative;
-    display: inline-block;
-    width: 100%;
-    margin-right: 1rem;
-  }
-  
-  .select::after {
-    content: "";
-    position: absolute;
-    top: 50%;
-    right: 0;
-    transform: translateY(-50%);
-    width: 0;
-    height: 0;
-    border-top: 5px solid #333;
-    border-right: 5px solid transparent;
-    border-left: 5px solid transparent;
-    pointer-events: none;
-  }
-  
-  .select select {
-    width: 100%;
-    padding: 0.5rem;
-    font-size: 1rem;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    appearance: none;
-    background-color: #fff;
-  }
-  
-  .input {
-    width: 100%;
-    padding: 0.5rem;
-    font-size: 1rem;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    background-color: #fff;
-  }
-  
-  .button {
-    padding: 0.5rem 1rem;
-    font-size: 1rem;
-    font-weight: bold;
-    color: #fff;
-    background-color: #333;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .button:hover {
-    background-color: #555;
-  }
