+++ src/assets/locales/de.json
@@ -2,5 +2,6 @@
     "select_language": "Wähle Sprache",
     "select_font_size": "Wähle Schriftgröße",
     "chapter": "Kapitel",
-    "verse": "Vers"
+    "verse": "Vers",
+    "search": "Suche kommt bald!"
 }
