+++ src/assets/locales/en.json
@@ -2,5 +2,6 @@
     "select_language": "Select Language",
     "select_font_size": "Select Font Size",
     "chapter": "Chapter",
-    "verse": "Verse"
+    "verse": "Verse",
+    "search": "Search coming soon!"
 }
